requirements.txt
aiosmtplib==3.0.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgi-lifespan==2.1.0
asyncpg==0.30.0
bcrypt==4.0.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.2
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-mail==1.4.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.16
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.0.0
rich==14.0.0
rich-toolkit==0.14.3
rsa==4.9.1
sendgrid==6.12.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
standard==1.0.3
starlette==0.46.2
termcolor==3.1.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
